[UPDATE] Moved the gameplay functions to the gameplay.h to clear main file + implemented working user turns
